{{ cookiecutter.repo_name }}/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
{% if cookiecutter.python_interpreter != 'python3' %}

# backwards compatibility
pathlib2
{% endif %}

# package specific modules
pyyaml